create ClientType and ServerType protocols
for WebSocket and HTTP servers
