+++ Source/Server.swift
@@ -9,46 +9,12 @@
 import SwiftFoundation
 import CoreModel
 
+public struct Server { }
+
 /// This class will broadcast a managed object context over the network.
-public final class Server {
+public protocol ServerType: class {
     
-    // MARK: - Properties
     
-    public var dataSource: ServerDataSource
-    
-    public var delegate: ServerDelegate?
-    
-    public var permissionsDelegate: ServerPermissionsDelegate?
-    
-    /// Function for logging purposes
-    public var log: ((String) -> ())?
-    
-    /// Function for sending the WebSocket response
-    public var sendMessage: String -> () = {
-        
-        // for websocketd
-        print($0)
-        fflush(__stdoutp)
-    }
-    
-    // MARK: - Initialization
-    
-    public init(dataSource: ServerDataSource,
-        delegate: ServerDelegate? = nil,
-        permissionsDelegate: ServerPermissionsDelegate?) {
-        
-        self.dataSource = dataSource
-        self.delegate = delegate
-        self.permissionsDelegate = permissionsDelegate
-    }
-    
-    // MARK: - Methods
-    
-    /// Process string as input.
-    public func input(input: String) {
-        
-        
-    }
 }
 
 // MARK: - Supporting Classes
